implemented a sell-bonus for each sold item. Bonus >= 1 added as integer.

+++ trade/app/models/store/system_user.rb
@@ -102,7 +102,7 @@ module Store
       end
 
       seller.release_item(item)
-      seller.credits += item.price + item.price * SELL_BONUS
+      seller.credits += item.price + Integer(item.price * SELL_BONUS)
 
       item.deactivate
